Separated the temoa_config.py file into parts
removed the utility function to create .dat files from the config file for clarity
